Move config files to config directory. Improve naming.
* This commit moves `AppConfig.js` to `/config` directory.
* Former `config` object is renamed to `i18n` and content is pulled out
into its own file for better maintainability.
* Rename/improve naming withing i18n object to better reflect what the content is doing
* Update the gulp imports and html task accordingly
