+++ config/i18n.js
@@ -1,0 +1,24 @@
+export const i18n = {
+  en: {
+    lang: `en`,
+    country: `US`,
+    dir: `ltr`,
+    routes: [
+      {view: `home`, path: `/home`},
+      {view: `example`, path: `/example`},
+      {tpl: `donate`, name: `donate`},
+      {tpl: `programs`, name: `programs`},
+    ],
+  },
+  es: {
+    lang: `es`,
+    country: `DO`,
+    dir: `ltr`,
+    routes: [
+      {view: `home`, path: `/inicio`},
+      {view: `example`, path: `/ejemplo`},
+      {tpl: `donate`, name: `donar`},
+      {tpl: `programs`, name: `programas`},
+    ],
+  },
+};
